Refactor mvp folders
