+++ app/src/main/java/mx/yellowme/app/ListFragment.java
@@ -1,12 +1,24 @@
 package mx.yellowme.app;
 
-import android.support.v4.app.Fragment;
+import android.view.View;
+
+import mx.yellowme.waterproof.WaterproofFragment;
 
 
-public class ListFragment extends Fragment {
+public class ListFragment
+        extends WaterproofFragment {
 
     public ListFragment() {
         // Required empty public constructor
     }
 
+    @Override
+    public int getMainLayoutResource() {
+        return 0;
+    }
+
+    @Override
+    public void bindViews(View root) {
+
+    }
 }
