--- waterproof/src/main/java/mx/yellowme/waterproof/lists/DataSource.java
@@ -1,5 +1,0 @@
-package mx.yellowme.waterproof.lists;
-
-public interface DataSource<Model> {
-
-}
