--- waterproof/src/main/java/mx/yellowme/waterproof/lists/Deletable.java
@@ -1,5 +1,0 @@
-package mx.yellowme.waterproof.lists;
-
-public interface Deletable {
-
-}
