+++ waterproof/src/main/java/mx/yellowme/waterproof/mvp/lists/Deletable.java
@@ -1,0 +1,5 @@
+package mx.yellowme.waterproof.mvp.lists;
+
+public interface Deletable {
+
+}
