+++ waterproof/src/main/java/mx/yellowme/waterproof/mvp/lists/ItemListener.java
@@ -1,4 +1,4 @@
-package mx.yellowme.waterproof.lists;
+package mx.yellowme.waterproof.mvp.lists;
 
 public interface ItemListener<Item> {
     void onItemClick(Item clickedUser);
