+++ waterproof/src/main/java/mx/yellowme/waterproof/mvp/lists/Searchable.java
@@ -1,4 +1,4 @@
-package mx.yellowme.waterproof.lists;
+package mx.yellowme.waterproof.mvp.lists;
 
 public interface Searchable {
     void search(final String criteria, boolean onRemoteSource);
